# server/swagger.yaml
openapi: 3.0.3
info:
  title: Rocko Backend API Docs
  version: 1.0.0
servers:
  - url: https://backend.dev.rocko.cloud
  - url: http://localhost:5000
  - url: https://backend.prod.rocko.cloud
...
paths:
...
  /average_apr:
    get:
        tags:
          - asset
        summary: ""
        description: ""
        parameters:
          - name: openDate
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: month
          - name: network
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: sepolia
        responses:
          '200':
            description: ""
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    average_apr:
                      type: number
          '400':
            description: Invalid value
  /average_reward_rate:
    get:
        tags:
          - asset
        summary: ""
        description: ""
        parameters:
          - name: network
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: sepolia
        responses:
          '200':
            description: ""
            content:
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    properties:
                      average_reward_rate:
                        type: number
          '400':
            description: Invalid status value
  /reward_rate:
    get:
        tags:
          - asset
        summary: ""
        description: ""
        parameters:
          - name: network
            in: query
            description: ""
            required: false
            explode: true
            schema:
              type : string
              example: sepolia
        responses:
          '200':
            description: ""
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    borrow_reward_rate:
                      type: number
          '400':
            description: Invalid status value
...
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 